+++ src/scanoss/components.py
@@ -24,7 +24,7 @@
 import json
 import os
 import sys
-from typing import TextIO
+from typing import TextIO, Optional, List
 
 from pypac.parser import PACFile
 
@@ -62,13 +62,13 @@ class Components(ScanossBase):
                                     ver_details=ver_details, ca_cert=ca_cert, proxy=proxy, pac=pac,
                                     grpc_proxy=grpc_proxy, timeout=timeout)
 
-    def load_purls(self, json_file: str = None, purls: [] = None) -> dict:
+    def load_purls(self, json_file: Optional[str] = None, purls: Optional[List[str]] = None) -> Optional[dict]:
         """
         Load the specified purls and return a dictionary
 
         :param json_file: JSON PURL file (optional)
         :param purls: list of PURLs (optional)
-        :return: PURL Request dictionary
+        :return: PURL Request dictionary or None
         """
         if json_file:
             if not os.path.isfile(json_file) or not os.access(json_file, os.R_OK):
@@ -81,6 +81,9 @@ class Components(ScanossBase):
                     self.print_stderr(f'ERROR: Problem parsing input JSON: {e}')
                     return None
         elif purls:
+            if not all(isinstance(purl, str) for purl in purls):
+                self.print_stderr('ERROR: PURLs must be a list of strings.')
+                return None
             parsed_purls = []
             for p in purls:
                 parsed_purls.append({'purl': p})
@@ -295,6 +298,32 @@ class Components(ScanossBase):
             return False
         self.print_msg('Sending PURLs to Component Versions API...')
         response = self.grpc_api.get_component_versions_json(request)
+        if response:
+            print(json.dumps(response, indent=2, sort_keys=True), file=file)
+            success = True
+            if output_file:
+                self.print_msg(f'Results written to: {output_file}')
+        self._close_file(output_file, file)
+        return success
+
+    def get_provenance_details(self, json_file: str = None, purls: [] = None, output_file: str = None) -> bool:
+        """
+        Retrieve the semgrep details for the supplied PURLs
+
+        :param json_file: PURL JSON request file (optional)
+        :param purls: PURL request array (optional)
+        :param output_file: output filename (optional). Default: STDOUT
+        :return: True on success, False otherwise
+        """
+        success = False
+        purls_request = self.load_purls(json_file, purls)
+        if purls_request is None or len(purls_request) == 0:
+            return False
+        file = self._open_file_or_sdtout(output_file)
+        if file is None:
+            return False
+        self.print_msg('Sending PURLs to Provenance API for decoration...')
+        response = self.grpc_api.get_provenance_json(purls_request)
         if response:
             print(json.dumps(response, indent=2, sort_keys=True), file=file)
             success = True